Noise::write now reports the correct amount of bytes (#99)
* Noise::write now reports the correct amount of bytes
* Fix data handling in Noise::read() & Noise::readSome()

+++ src/crypto/key_marshaller/key_marshaller_impl.cpp
@@ -84,7 +84,7 @@ namespace libp2p::crypto::marshaller {
       const ProtobufKey &proto_key) const {
     protobuf::PublicKey protobuf_key;
     if (!protobuf_key.ParseFromArray(proto_key.key.data(),
-                                     proto_key.key.size())) {
+                                     static_cast<int>(proto_key.key.size()))) {
       return CryptoProviderError::FAILED_UNMARSHAL_DATA;
     }
 
@@ -101,7 +101,7 @@ namespace libp2p::crypto::marshaller {
       const ProtobufKey &proto_key) const {
     protobuf::PublicKey protobuf_key;
     if (!protobuf_key.ParseFromArray(proto_key.key.data(),
-                                     proto_key.key.size())) {
+                                     static_cast<int>(proto_key.key.size()))) {
       return CryptoProviderError::FAILED_UNMARSHAL_DATA;
     }
